refactor(codegen): extract lower_locus_instantiation into locus/instantiation.rs (Round 4d)
1 fn lifted into LocusInstantiate<'ctx>: lower_locus_instantiation
(~2,676 lines — the heaviest single function in the codebase). This
is the locus constructor path:

- arena acquire (Fresh / Subregion / RecpoolFixed / RecpoolSlab)
- capacity slot allocation + form-specific init (vec / hashmap /
  ring_buffer / shm_ring) — most of the inline form codegen that
  the original plan was going to extract in Round 2 actually lives
  here
- param init, recursive child instantiation
- bus subscriber registration (legacy queue + shm_ring + keyed)
- closure-eval scheduling (birth / tick / duration synthesis)
- recovery-handler routing (resolve_failure_route + parent slot
  threading)
- pinned-thread spawn + cooperative-pool registration
- deferred-dissolve frame for long-lived loci (those with bus
  subscribers, dissolve closures, etc.)

Surface bumps to pub(crate):
- Cx fields: program, current_arena_override, params_init_self,
  defer_next_locus_dissolve, instantiating_for_parent_field,
  instantiating_into_payload_arena,
  placement_for_next_locus_instantiation,
  cooperative_pool_for_next_locus_instantiation,
  coop_pool_run_wrappers
- Cx methods: wrap_literal_as_view, const_param,
  zero_accumulator_slot, locus_satisfies_interface,
  coerce_to_interface, alloca_in_entry_with_nulled_arena
- Free fn: chunk_hint_for_coop_pool

codegen.rs: 31358 -> 28683 (-2675 lines). Biggest single-sub-round
drop yet.
